Improvement on fragment classes handling
Defining empty methods for methods that are supposed to be overriden in the child classes should be avoided.
This will cause a confusion for understanding the flow of the class and it doesn't make clear that those methods are supposed to be overriden.
It is possible that the child class doesn't override that method and you'll spend time to debug wondering why the fragment doesn't work as it should.

Changes:
* Make RecyclerViewFragment class abstract so that child classes must override the abstract methods
  This will guarantee that child classes won't forget to implement important methods because java compiler will warn us with errors.
* Don't use a common method name that could conflict with base/super classes
* Refactor out the AsyncTask and pass the savedInstanceState

Change-Id: I8791ca630c427286b1ecddc9b8f8ee85a5694ebd